Refactor matrix logic to reduce import load
